+++ Sources/EudiWalletKit/Services/OpenId4VpService.swift
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 
-Created on 04/10/2023 
+Created on 04/10/2023
 */
 
 import Foundation
@@ -33,20 +33,25 @@ import X509
 public final class OpenId4VpService: @unchecked Sendable, PresentationService {
 	public var status: TransferStatus = .initialized
 	var openid4VPlink: String
-	// map of document id to data format
+	// map of document-id to data format
 	var dataFormats: [String: DocDataFormat]!
-	// map of document id to data
+	// map of document-id to data
 	var docs: [String: Data]!
 	var docDisplayNames: [String: [String: [String: String]]?]!
-	// map of document id to IssuerSigned
+	// map of document-id to IssuerSigned
 	var docsCbor: [String: IssuerSigned]!
 	/// map of document id to document type
 	var idsToDocTypes: [String: String]!
-	// map of document id to SignedSDJWT
+	// map of document-id to SignedSDJWT
 	var docsSdJwt: [String: SignedSDJWT]!
+	// map of document-id to hashing algorithm
+	var docsHashingAlgs: [String: String]!
+	/// IACA root certificates
 	var iaca: [SecCertificate]!
 	// map of docType to data format (formats requested)
 	var formatsRequested: [String: DocDataFormat]!
+	/// map of docType to inputDescriptor-id
+	var inputDescriptorMap: [String: String]!
 	var dauthMethod: DeviceAuthMethod
 	var devicePrivateKeys: [String: CoseKeyPrivate]!
 	var logger = Logger(label: "OpenId4VpService")
@@ -59,6 +64,7 @@ public final class OpenId4VpService: @unchecked Sendable, PresentationService {
 	var readerCertificateIssuer: String?
 	var readerCertificateValidationMessage: String?
 	var vpNonce: String!
+	var vpClientId: String!
 	var mdocGeneratedNonce: String!
 	var sessionTranscript: SessionTranscript!
 	var eReaderPub: CoseKey?
@@ -71,7 +77,9 @@ public final class OpenId4VpService: @unchecked Sendable, PresentationService {
 		let objs = parameters.toInitializeTransferInfo()
 		dataFormats = objs.dataFormats; docs = objs.documentObjects; devicePrivateKeys = objs.privateKeyObjects
 		iaca = objs.iaca; dauthMethod = objs.deviceAuthMethod
-		idsToDocTypes = objs.idsToDocTypes; docDisplayNames = objs.docDisplayNames
+		idsToDocTypes = objs.idsToDocTypes
+		docDisplayNames = objs.docDisplayNames
+		docsHashingAlgs = objs.hashingAlgs
 		guard let openid4VPlink = String(data: qrCode, encoding: .utf8) else {
 			throw PresentationSession.makeError(str: "QR_DATA_MALFORMED")
 		}
@@ -80,7 +88,7 @@ public final class OpenId4VpService: @unchecked Sendable, PresentationService {
 		self.openId4VpVerifierLegalName = openId4VpVerifierLegalName
 		self.urlSession = urlSession
 	}
-	
+
 	public func startQrEngagement(secureAreaName: String?, crv: CoseEcCurve) async throws -> String {
 		if unlockData == nil {
 			unlockData = [String: Data]()
@@ -91,7 +99,7 @@ public final class OpenId4VpService: @unchecked Sendable, PresentationService {
 		}
 		return ""
 	}
-	
+
 	///  Receive request from an openid4vp URL
 	///
 	/// - Returns: The requested items.
@@ -110,14 +118,14 @@ public final class OpenId4VpService: @unchecked Sendable, PresentationService {
 						eReaderPub = CoseKey(x: [UInt8](xd), y: [UInt8](yd), crv: crvType)
 					}
 					let responseUri = if case .directPostJWT(let uri) = vp.responseMode { uri.absoluteString } else { "" }
-					vpNonce = vp.nonce
+					vpNonce = vp.nonce; vpClientId = vp.client.id.originalClientId
 					mdocGeneratedNonce = Openid4VpUtils.generateMdocGeneratedNonce()
-					sessionTranscript = Openid4VpUtils.generateSessionTranscript(clientId: vp.client.id,
+					sessionTranscript = Openid4VpUtils.generateSessionTranscript(clientId: vp.client.id.originalClientId,
 						responseUri: responseUri, nonce: vp.nonce, mdocGeneratedNonce: mdocGeneratedNonce)
 					logger.info("Session Transcript: \(sessionTranscript.encode().toHexString()), for clientId: \(vp.client.id), responseUri: \(responseUri), nonce: \(vp.nonce), mdocGeneratedNonce: \(mdocGeneratedNonce!)")
 					self.presentationDefinition = vp.presentationDefinition
-					let (items, fmtsReq) = try Openid4VpUtils.parsePresentationDefinition(vp.presentationDefinition, idsToDocTypes: idsToDocTypes, dataFormats: dataFormats, docDisplayNames: docDisplayNames, logger: logger)
-					self.formatsRequested = fmtsReq
+					let (items, fmtsReq, imap) = try Openid4VpUtils.parsePresentationDefinition(vp.presentationDefinition, idsToDocTypes: idsToDocTypes, dataFormats: dataFormats, docDisplayNames: docDisplayNames, logger: logger)
+					self.formatsRequested = fmtsReq; self.inputDescriptorMap = imap
 					guard let items else { throw PresentationSession.makeError(str: "Invalid presentation definition") }
 					var result = UserRequestInfo(docDataFormats: fmtsReq, validItemsRequested: items)
 					logger.info("Verifer requested items: \(items.mapValues { $0.mapValues { ar in ar.map(\.elementIdentifier) } })")
@@ -132,13 +140,13 @@ public final class OpenId4VpService: @unchecked Sendable, PresentationService {
 				}
 			}
 	}
-	
+
 	fileprivate func makeCborDocs() {
 		docsCbor = docs.filter { k,v in Self.filterFormat(dataFormats[k]!, fmt: .cbor)} .mapValues { IssuerSigned(data: $0.bytes) }.compactMapValues { $0 }
 	}
 
 	func generateCborVpToken(itemsToSend: RequestItems) async throws -> VpToken.VerifiablePresentation {
-		let resp = try await MdocHelpers.getDeviceResponseToSend(deviceRequest: nil, issuerSigned: docsCbor, docDisplayNames: docDisplayNames, selectedItems: itemsToSend, eReaderKey: eReaderPub, devicePrivateKeys: devicePrivateKeys, sessionTranscript: sessionTranscript, dauthMethod: .deviceSignature, unlockData: unlockData) 
+		let resp = try await MdocHelpers.getDeviceResponseToSend(deviceRequest: nil, issuerSigned: docsCbor, docDisplayNames: docDisplayNames, selectedItems: itemsToSend, eReaderKey: eReaderPub, devicePrivateKeys: devicePrivateKeys, sessionTranscript: sessionTranscript, dauthMethod: .deviceSignature, unlockData: unlockData)
 		guard let resp else { throw PresentationSession.makeError(str: "DOCUMENT_ERROR") }
 		let vpTokenStr = Data(resp.deviceResponse.toCBOR(options: CBOROptions()).encode()).base64URLEncodedString()
 		return VpToken.VerifiablePresentation.msoMdoc(vpTokenStr)
@@ -159,10 +167,10 @@ public final class OpenId4VpService: @unchecked Sendable, PresentationService {
 		}
 		logger.info("Openid4vp request items: \(itemsToSend.mapValues { $0.mapValues { ar in ar.map(\.elementIdentifier) } })")
 		if unlockData == nil { _ = try await startQrEngagement(secureAreaName: nil, crv: .P256) }
-		if formatsRequested.allSatisfy({ (key: String, value: DocDataFormat) in value == .cbor }) {
+		if formatsRequested.count == 1, formatsRequested.allSatisfy({ (key: String, value: DocDataFormat) in value == .cbor }) {
 			makeCborDocs()
 			let vpToken = try await generateCborVpToken(itemsToSend: itemsToSend)
-			try await SendVpToken([vpToken], pd, resolved, onSuccess)
+			try await SendVpToken([(pd.inputDescriptors.first!.id, vpToken)], pd, resolved, onSuccess)
 		} else {
 			if formatsRequested.first(where: { (key: String, value: DocDataFormat) in value == .cbor }) != nil {
 				makeCborDocs()
@@ -170,35 +178,41 @@ public final class OpenId4VpService: @unchecked Sendable, PresentationService {
 			let parser = CompactParser()
 			let docStrings = docs.filter { k,v in Self.filterFormat(dataFormats[k]!, fmt: .sdjwt)}.compactMapValues { String(data: $0, encoding: .utf8) }
 			docsSdJwt = docStrings.compactMapValues { try? parser.getSignedSdJwt(serialisedString: $0) }
-			var presentations = [VpToken.VerifiablePresentation]()
+			var inputToPresentations = [(String, VpToken.VerifiablePresentation)]()
 			// support sd-jwt documents
 			for (docId, nsItems) in itemsToSend {
+				guard let docType = idsToDocTypes[docId], let inputDescrId = inputDescriptorMap[docType] else { continue }
 				if dataFormats[docId] == .cbor {
-					//continue // cannot put mixed dataFormats for now
 					if docsCbor == nil { makeCborDocs() }
 					let itemsToSend1 = Dictionary(uniqueKeysWithValues: [(docId, nsItems)])
 					let vpToken = try await generateCborVpToken(itemsToSend: itemsToSend1)
-					 presentations.append(vpToken)
+					 inputToPresentations.append((inputDescrId, vpToken))
 				} else if dataFormats[docId] == .sdjwt {
 					let docSigned = docsSdJwt[docId]; let dpk = devicePrivateKeys[docId]
-					guard let docSigned, let dpk, let items = nsItems.first?.value.map(\.elementIdentifier) else { continue }
+					guard let docSigned, let dpk, let items = nsItems.first?.value else { continue }
 					let unlockData = try await dpk.secureArea.unlockKey(id: docId)
 					let keyInfo = try await dpk.secureArea.getKeyInfo(id: docId);	let dsa = keyInfo.publicKey.crv.defaultSigningAlgorithm
 					let signer = try SecureAreaSigner(secureArea: dpk.secureArea, id: docId, ecAlgorithm: dsa, unlockData: unlockData)
 					let signAlg = try SecureAreaSigner.getSigningAlgorithm(dsa)
-					guard let presented = try await Openid4VpUtils.getSdJwtPresentation(docSigned, signer: signer, signAlg: signAlg, requestItems: items, nonce: vpNonce, aud: openid4VPlink) else { continue }
-					presentations.append(VpToken.VerifiablePresentation.generic(presented.serialisation))
+					let hai = HashingAlgorithmIdentifier(rawValue: docsHashingAlgs[docId] ?? "") ?? .SHA3256
+					guard let presented = try await Openid4VpUtils.getSdJwtPresentation(docSigned, hashingAlg: hai.hashingAlgorithm(), signer: signer, signAlg: signAlg, requestItems: items, nonce: vpNonce, aud: vpClientId) else {
+						continue
+					}
+					inputToPresentations.append((inputDescrId, VpToken.VerifiablePresentation.generic(presented.serialisation)))
 				}
 			}
-			try await SendVpToken(presentations, pd, resolved, onSuccess)
+			try await SendVpToken(inputToPresentations, pd, resolved, onSuccess)
 		}
 	}
 	/// Filter document accordind to the raw format value
 	static func filterFormat(_ df: DocDataFormat, fmt: DocDataFormat) -> Bool { df == fmt }
 
-	fileprivate func SendVpToken(_ vpTokens: [VpToken.VerifiablePresentation]?, _ pd: PresentationDefinition, _ resolved: ResolvedRequestData, _ onSuccess: ((URL?) -> Void)?) async throws {
+	fileprivate func SendVpToken(_ vpTokens: [(String, VpToken.VerifiablePresentation)]?, _ pd: PresentationDefinition, _ resolved: ResolvedRequestData, _ onSuccess: ((URL?) -> Void)?) async throws {
 		let consent: ClientConsent = if let vpTokens {
-			.vpToken(vpToken: .init(apu: mdocGeneratedNonce.base64urlEncode, verifiablePresentations: vpTokens), presentationSubmission: .init(id: UUID().uuidString, definitionID: pd.id, descriptorMap: pd.inputDescriptors.map {DescriptorMap(id: $0.id, format: $0.formatContainer?.formats.first?["designation"].string ?? "", path: "$")}))
+			.vpToken(vpToken: .init(apu: mdocGeneratedNonce.base64urlEncode, verifiablePresentations: vpTokens.map(\.1)), presentationSubmission: .init(id: UUID().uuidString, definitionID: pd.id, descriptorMap: vpTokens.enumerated().map { i,v in
+			 let descr = pd.inputDescriptors.first(where: { $0.id == v.0 })!
+			 return DescriptorMap(id: descr.id, format: descr.formatContainer?.formats.first?["designation"].string ?? "", path: vpTokens.count == 1 ? "$" : "$[\(i)]")
+			}))
 		} else { .negative(message: "Rejected") }
 		// Generate a direct post authorisation response
 		let response = try AuthorizationResponse(resolvedRequest: resolved, consent: consent, walletOpenId4VPConfig: getWalletConf(verifierApiUrl: openId4VpVerifierApiUri, verifierLegalName: openId4VpVerifierLegalName))
@@ -211,7 +225,7 @@ public final class OpenId4VpService: @unchecked Sendable, PresentationService {
 			throw PresentationSession.makeError(str: reason)
 		}
 	}
-	
+
 	lazy var chainVerifier: CertificateTrust = { [weak self] certificates in
 		guard let self else { return false }
 		let chainVerifier = eudi_lib_sdjwt_swift.X509CertificateChainVerifier()
@@ -227,7 +241,7 @@ public final class OpenId4VpService: @unchecked Sendable, PresentationService {
 		self.readerCertificateValidationMessage = validationMessages.joined(separator: "\n")
 		return result
 	}
-	
+
 	/// OpenId4VP wallet configuration
 	func getWalletConf(verifierApiUrl: String?, verifierLegalName: String?) -> SiopOpenId4VPConfiguration? {
 		guard let rsaPrivateKey = try? KeyController.generateRSAPrivateKey(), let privateKey = try? KeyController.generateECDHPrivateKey(),
@@ -242,6 +256,6 @@ public final class OpenId4VpService: @unchecked Sendable, PresentationService {
 		let res = SiopOpenId4VPConfiguration(subjectSyntaxTypesSupported: [.decentralizedIdentifier, .jwkThumbprint], preferredSubjectSyntaxType: .jwkThumbprint, decentralizedIdentifier: try! DecentralizedIdentifier(rawValue: "did:example:123"), signingKey: privateKey, signingKeySet: keySet, supportedClientIdSchemes: supportedClientIdSchemes, vpFormatsSupported: [], session: urlSession)
 		return res
 	}
-	
+
 }